Add a wrapper for Elmo objects and convert all classes to use this wrapper instead.
Simal code should not touch the Elmo DOAP class, instead they should do all work via these wrappers.

The wrapper is required in order to provide easy access within the wicket framework, for example the Elmo class returns a Set of names, which, whilst correct makes it difficult to render well in the UI. This wrapper allows us to control which name will be returned.