Hopefully this will resolve the problem with the CI build.
The issue appears to have been that <simal:personID> is only guarenteed unique to a single repository. Thus we cannot be sure that the ID is available when loading from an RDF/XML file. 

This showed up in the CI server as it was cleaning everything between builds and thus removing the local properties file, which houses the next ID numbers. This rather large commit does many things, including:

- create a properties file for use in tests. This file is deleted at the start of the test runs

- when loading from RDF/XML try and respect the <simal:projectID> and <simal:personID> values. If they can't be used then throw away the suggested values and assign new ones.

- fix ISSUE 88 (project ids not being written to the repo)

- remove transaction support from the repository and related tests, it was causing all sorts of problems.

- make tests responsible for putting test data back into a consistent state

- various refactorings