A branch for working on open social integration.
A branch is necessary as some of the functionality requires a patch to be applied to openSocial and I don't have the commit rights to do that. Creating a branch will allow me to continue work on required features in head whilst waiting for the open social patch to be applied.