Resources now get a world unique simalID. That is, each running instance of Simal has a world unique ID and each Project, Category and Person within that instance gets a unique ID. These two things together give us a world unique identifier.
If a document is imported from another instance of Simal this will mean the simalID is not valid locally. We therefore add a seeAlso tuple for this id and assign a new ID for this instance of SImal.

In doing this I discovered quite a few weaknesses in the way we were handling data imports. So they've been fixed in the process.

I've also hardened up the tests considerably (as well has having to refactor many becase of the change in the way the identifiers work.