- Added schema documentation to the wiki. - Added test review entry to testDOAP.xml - Added IReview interface for working with reviews - Added Review class for working with reviews - Aded number of revews to summary on tools page - Added review list panel - Added abilty to add a review for a project from the project detail page
Openness rating
========

We need a number between 0 and 100 to indicate how open we believe a project to be. This should only be calculated on projects that we have manually inspected. It is calculated each time the value is requested.

We therefore need a record of whether or not we ave inspected a project entry. We will use:

<simal:Review rdf:about="http://foo.org/project#bar" simal:when="DATE_STRING" simal:reviewer="http://foo.org/#me" simal:type="openness"/>

Then we need a method in the project class:

/**
  * Get a value between 0 and 100 that represents the openness of the project development model.
  * 0 is low, 100 is high.
  */
public int getOpennessRating();
  
In the first instance this is a very basic calculation that looks to see if a project has the infrastructure of an open development project (i.e. Mailing list, version control, issue tracker etc.) and calculates a percentage based on how many of these are present.
  
We also report when the project was last reviewed on the project summary page.

New Task
Move the logic for calculating openness into a new module specifically for the task. Use the Openness document written by Pia to calculate the rating. 
Priority: Low