We no longer allow doap:* values inside of the simal:Project elements. This makes filtering items easier. Simal:Project now points to doap:Project files using rdf:seeAlso. This solves the problem where projects were sometimes returned twice.
This was implemented because it was proving to be very difficult to filter for projects with/without certain properties.

Added some methods and tests to ensure that this works OK.

Improved process for identifying duplicates and refactored ingest of projects.

Because of the change in the use of simal:Project we need to update the way PIMS data is imported. We now create an RDF/XML document and use the standard import chain. This makes the whole process more maintainable.