Add links from the report page to lists of projects with/without given properties.
I've had to disable some of the tests as a result of this, I just can't seem to make the tests work (not finding the elements in the page). It's too late for me to fix now and since the percentage test is still present I figured I was safe commiting this now as other tests will be failing as a result of a bug fix in core that is ensuring duplicate projects are not being returned.