Add support for a Wookie Server and add some test features to the ToolsPage.
Currently we list all available widgets on the ToolsPage. Alongside each widget we have an "instantiate" link. This takes us to a page showing us the widget in action.

Returning to the ToolsPage will now show this (and any other previously instantiated widgets).

This work allows us to expose features of Simal as Widgets, thus making it easy to embed Simal into other applications.

Note that, at present, there is no support for different users or per application API keys so this is not ready for the big time yet (per instance widgets, multi-user widgets, etc.)