A simple project list widget. It is intended to replace the current project browser panel in some use cases.
All this does at present is retrieve the projects list via the REST Api and display searchable list of the projects. Clicking on a project in the list opens a simple project details page.

It's not a production ready widget yet, but it is a good start. Things that still need doing include:

- create a setting for the location of the Wookie instance (its currently hard coded as localhost:8888/wookie)

- address performance issues (currently all projects are loaded and all detail pages are created on initialisation)

- add more detail to the detail page