This is a potentially risky commit with respect to side effects. I'd appreciate a code review, all tests pass but that's not enough.
This commit removes the toJSON(boolean) method. Normally I would deprecate it, but toJSON(false) was never called, other than in a test. I strongly suspect this was a throwback to some previous issue that has now gone away.

I've removed it to simplify the JSON workflow, though we really ought to be using a standard JSON library not building the JSON by hand. See ISSUE 73